% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{accept_reject}
\alias{accept_reject}
\title{Performs Metropolis accept-reject step}
\usage{
accept_reject(
  logp_current,
  logp_proposed,
  current_parameters,
  proposed_parameters
)
}
\arguments{
\item{logp_current}{current log-probability value}

\item{logp_proposed}{proposed log-probability value}

\item{current_parameters}{current parameter(s) value}

\item{proposed_parameters}{proposed parameter(s) value}
}
\value{
a named list with two elements: 'parameter', a parameter value (which
may be a non-scalar); and 'logp', the log-probability corresponding to the
parameter value returned
}
\description{
Performs Metropolis accept-reject step
}
