% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{case_undetected_prob}
\alias{case_undetected_prob}
\title{Probability case remains undetected over time}
\usage{
case_undetected_prob(t_primed, t, delay_mean, delay_sd)
}
\arguments{
\item{t_primed}{time subsequent to time when case originates}

\item{t}{time when case originates}

\item{delay_mean}{mean of gamma distribution governing reporting delay distribution}

\item{delay_sd}{sd of gamma distribution governing reporting delay distribution}
}
\value{
a probability
}
\description{
Probability case remains undetected over time
}
