% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{combine_chains}
\alias{combine_chains}
\title{Combines Markov chains across multiple runs of mcmc_single}
\usage{
combine_chains(list_of_results)
}
\value{
a named list of three tibbles: "cases", "Rt" and "reporting" which
contain estimates of the model parameters with chain index  included
}
\description{
Combines Markov chains across multiple runs of mcmc_single
}
