% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{construct_w_matrix}
\alias{construct_w_matrix}
\title{Constructs a matrix of w vectors}
\usage{
construct_w_matrix(w, piece_width)
}
\arguments{
\item{w}{a vector of weights}

\item{piece_width}{the width of an Rt piece}
}
\value{
a matrix of ws
}
\description{
Constructs a matrix of w vectors
}
