% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_checking.R
\name{convert_df_to_posterior_format}
\alias{convert_df_to_posterior_format}
\title{Converts a single tibble to posterior format}
\usage{
convert_df_to_posterior_format(
  res_df,
  parameter_cols,
  names_cols,
  names_prefix
)
}
\arguments{
\item{res_df}{one of the three tibbles returned by running mcmc}

\item{parameter_cols}{the column(s) containing the parameter draws}

\item{names_cols}{another column specifying an index for parameter draws}

\item{names_prefix}{the prefix for naming the columns of the output tibble}
}
\value{
a tibble in the format required by posterior package
}
\description{
Converts a single tibble to posterior format
}
