% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergence_checking.R
\name{convert_results_to_posterior_format}
\alias{convert_results_to_posterior_format}
\title{Converts results to format required for posterior package}
\usage{
convert_results_to_posterior_format(results)
}
\arguments{
\item{results}{an object returned by running mcmc}
}
\value{
a named list of three tibbles: "cases", "Rt" and "reporting" which contain estimates of the model parameters
}
\description{
Converts results to format required for posterior package
}
