% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{detected_after_unobserved_prob}
\alias{detected_after_unobserved_prob}
\title{Probability a case is detected by a later time given it was unobserved
earlier}
\usage{
detected_after_unobserved_prob(day_2, day_1, day_onset, delay_parameters)
}
\arguments{
\item{day_2}{day of last observation: day_2 > day_1 >= day_onset}

\item{day_1}{day of first observation: day_1 >= day_onset}

\item{day_onset}{day when case originates}

\item{delay_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}
}
\value{
a probability
}
\description{
This is the probability that a case originating on day day_onset is detected
between days day_1 and day_2 given that it was unobserved on day day_1. This
probability is calculated using the survival functions (which indicate the
probability a case remains undetected by a certain time):
\deqn{prob(detect) = \frac{S(day_1|day_onset) - S(day_2|day_onset)}{S(day_1|day_onset)}}
}
