% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{dgamma_mean_sd}
\alias{dgamma_mean_sd}
\title{Reparameterisation of gamma pdf using mean and sd}
\usage{
dgamma_mean_sd(x, mu, sigma, ...)
}
\arguments{
\item{x}{a value to be evaluated at}

\item{mu}{a mean}

\item{sigma}{a standard deviation}

\item{...}{other arguments passed to method}
}
\value{
a density
}
\description{
Reparameterisation of gamma pdf using mean and sd
}
