% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expected_cases}
\alias{expected_cases}
\title{Gives cases expected given history of cases and Rt}
\usage{
expected_cases(Rt, weights, cases_history)
}
\arguments{
\item{Rt}{effective reproduction number}

\item{weights}{a vector of normalised weights}

\item{cases_history}{a vector containing history of cases arranged from
recent to past}
}
\value{
an expected number of cases
}
\description{
Gives cases expected given history of cases and Rt
}
