% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gamma_discrete_pmf}
\alias{gamma_discrete_pmf}
\title{Discrete gamma probability mass function}
\usage{
gamma_discrete_pmf(day, serial_parameters)
}
\arguments{
\item{day}{day to evaluate pmf}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}
}
\value{
a probability
}
\description{
Discrete gamma probability mass function
}
