% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{generate_snapshots}
\alias{generate_snapshots}
\title{Generates case data with snapshots of reported cases at each onset day}
\usage{
generate_snapshots(
  days_total,
  Rt_function,
  serial_parameters,
  reporting_parameters,
  thinned = FALSE,
  kappa = 1000,
  days_max_follow_up = 30,
  initial_parameters = list(mean = 30, length = 20),
  serial_max = 40
)
}
\arguments{
\item{days_total}{number of days to simulate process for (post-initial
period)}

\item{Rt_function}{takes day as an input and outputs an Rt value}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}

\item{reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}

\item{thinned}{a Boolean indicating whether to thin cases to only informative time points}

\item{kappa}{over-dispersion parameter of negative binomial}

\item{days_max_follow_up}{max days at which to simulate reporting case
trajectory}

\item{initial_parameters}{a named list of 'mean' and 'length'
which is used to generate seed cases by sampling from a negative binomial
distribution}

\item{serial_max}{maximum point at which to truncate sum in renewal process}
}
\value{
a tibble with observed case trajectories for each time of onset
}
\description{
Generates case data with snapshots of reported cases at each onset day
}
\examples{
library(incidenceinflation)
# generate case series for 100 days with time-varying Rt
days_total <- 100

# make Rt interpolation function
v_Rt <- c(rep(1.3, 25), rep(1, 25), rep(2, 50))
Rt_function <- stats::approxfun(1:days_total, v_Rt)

# serial interval parameters
s_params <- list(mean=5, sd=1)

# renewal over-dispersion parameter
kappa <- 2

# reporting delays
r_params <- list(mean=10, sd=5)
reported_cases <- generate_snapshots(
    days_total, Rt_function,
    s_params, r_params)
}
