% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ILI_2014}
\alias{ILI_2014}
\title{Influenza-like illness (ILI) in US from 2014-2015}
\format{
A data frame with 729 rows and 4 variables:
\describe{
\item{date_onset}{date when case onset}
\item{time_onset}{numeric time when case onset}
\item{time_reported}{time case reported}
\item{cases_reported}{number of cases reported}
}
}
\source{
\url{https://doi.org/10.1371/journal.pcbi.1007735}
}
\usage{
ILI_2014
}
\description{
A dataset containing ILI case data with reporting delay information
}
\keyword{datasets}
