% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{max_uncertain_days}
\alias{max_uncertain_days}
\title{Calculates max number of days we are uncertain about reporting}
\usage{
max_uncertain_days(p_gamma_cutoff, reporting_parameters)
}
\arguments{
\item{p_gamma_cutoff}{a p value (0 <= p <= 1) indicating the threshold above which
we deem certainty}

\item{reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}
}
\value{
a number of days
}
\description{
Calculates max number of days we are uncertain about reporting
}
