% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{maximise_Rt_reporting_overdispersion}
\alias{maximise_Rt_reporting_overdispersion}
\title{Maximises log-probability using L-BFGS-B}
\usage{
maximise_Rt_reporting_overdispersion(current, model)
}
\arguments{
\item{current}{a named list containing the current parameter values}

\item{model}{an initialised Stan model}
}
\value{
a named list with a single element: 'current'
}
\description{
Maximises log-probability using L-BFGS-B
}
