% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{mcmc}
\alias{mcmc}
\title{Runs MCMC or optimisation to estimate Rt, cases and reporting parameters}
\usage{
mcmc(
  niterations,
  snapshot_with_Rt_index_df,
  priors,
  serial_parameters,
  initial_reporting_parameters,
  initial_Rt,
  reporting_metropolis_parameters = list(mean_step = 0.25, sd_step = 0.1),
  serial_max = 40,
  p_gamma_cutoff = 0.99,
  maximise = FALSE,
  print_to_screen = TRUE
)
}
\arguments{
\item{niterations}{number of MCMC iterations to run}

\item{snapshot_with_Rt_index_df}{a tibble with
four columns: time_onset, time_reported, cases_reported, Rt_index}

\item{priors}{a named list with: 'Rt', 'reporting', 'max_cases'. These take
the form: 'Rt' is a named list with elements 'shape' and 'rate' describing
the gamma prior for each Rt; 'reporting' is a named list with elements
'mean_mu', 'mean_sigma', 'sd_mu', 'sd_sigma' representing the gamma
prior parameters for the mean and sd parameters of the reporting parameters
(itself described by a gamma distribution); max cases controls the upper
limit of the discrete uniform distribution representing the prior on true
cases}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}

\item{initial_reporting_parameters}{a list with two named elements: 'mean', 'sd'
indicating an initial guess of the mean and sd of the reporting delay distribution}

\item{initial_Rt}{initial guess of the Rt values in each of the piecewise segments.
Provided in the form of a tibble with columns: 'Rt_index' and 'Rt'}

\item{reporting_metropolis_parameters}{named list of 'mean_step', 'sd_step' containing
step sizes for Metropolis step}

\item{serial_max}{maximum point at which to truncate sum in renewal process}

\item{maximise}{whether to estimate MAP values of parameters (if true) or
sample parameter values using MCMC (if false). By default this is false.}
}
\value{
a named list of three tibbles: "cases", "Rt" and "reporting" which contain estimates of the model parameters
}
\description{
Runs MCMC or optimisation to estimate Rt, cases and reporting parameters
}
