% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{measles_NL_2013}
\alias{measles_NL_2013}
\title{Measles outbreak data in Netherlands from 2013-2014}
\format{
A data frame with 5934 rows and 4 variables:
\describe{
\item{date_onset}{date when case onset}
\item{time_onset}{numeric time when case onset}
\item{time_reported}{time case reported}
\item{cases_reported}{number of cases reported}
}
}
\source{
\url{https://github.com/kassteele/Nowcasting}
}
\usage{
measles_NL_2013
}
\description{
A dataset containing measles case data with reporting delay information
}
\keyword{datasets}
