% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{metropolis_step}
\alias{metropolis_step}
\title{Sample reporting parameters using a single Metropolis step}
\usage{
metropolis_step(
  snapshot_with_true_cases_df,
  current_reporting_parameters,
  prior_parameters,
  metropolis_parameters
)
}
\arguments{
\item{snapshot_with_true_cases_df}{a tibble with four columns:
time_onset, time_reported, cases_reported, cases_true}

\item{current_reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}

\item{prior_parameters}{named list with elements 'mean_mu', 'mean_sigma', 'sd_mu',
'sd_sigma' representing the gamma prior parameters for the mean and sd
parameters of the reporting parameters (itself described by a gamma
distribution)}

\item{metropolis_parameters}{named list of 'mean_step', 'sd_step' containing
step sizes for Metropolis step}
}
\value{
list of reporting parameters
}
\description{
Sample reporting parameters using a single Metropolis step
}
