% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{metropolis_step_overdispersion}
\alias{metropolis_step_overdispersion}
\title{Performs a single Metropolis step to update overdispersion parameter}
\usage{
metropolis_step_overdispersion(
  overdispersion_current,
  logp_current,
  cases_history_rt_df,
  serial_parameters,
  prior_overdispersion_parameter,
  overdispersion_metropolis_sd
)
}
\arguments{
\item{overdispersion_current}{current overdispersion parameter value (must exceed 0)}

\item{logp_current}{current log-probability value from previous update}

\item{cases_history_rt_df}{a tibble with three columns: 'time_onset',
'cases_true' and 'Rt'}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}
}
\value{
a named list with two elements: 'overdispersion' an overdispersion parameter
value and 'logp' the log-probability corresponding to the returned parameter
set
}
\description{
Performs a single Metropolis step to update overdispersion parameter
}
