% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{metropolis_step_Rt_reporting_overdispersion}
\alias{metropolis_step_Rt_reporting_overdispersion}
\title{Runs single Adaptive-Covariance Metropolis step}
\usage{
metropolis_step_Rt_reporting_overdispersion(
  current,
  model,
  mu,
  omega,
  log_lambda,
  eta,
  iteration
)
}
\arguments{
\item{current}{a named with with current parameter values}

\item{model}{a stan object generated by stan_initialisation}

\item{mu}{a tuned hyperparameter vector}

\item{omega}{a tuned hyperparameter covariance matrix}

\item{log_lambda}{a tuned hyperparameter value}

\item{eta}{a hyperparameter that when larger leads to slower decay in adaptations}

\item{iteration}{iteration count}
}
\value{
a named list with elements: 'current', 'log_p_current', 'mu', 'omega', 'log_lambda'
}
\description{
Runs single Adaptive-Covariance Metropolis step
}
