% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{metropolis_steps}
\alias{metropolis_steps}
\title{Sample reporting parameters using Metropolis MCMC}
\usage{
metropolis_steps(
  snapshot_with_true_cases_df,
  current_reporting_parameters,
  metropolis_parameters,
  ndraws
)
}
\arguments{
\item{snapshot_with_true_cases_df}{a tibble with four columns:
time_onset, time_reported, cases_reported, cases_true}

\item{current_reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}

\item{metropolis_parameters}{named list of 'mean_step', 'sd_step' containing
step sizes for Metropolis step}

\item{ndraws}{number of iterates of the Markov chain to simulate}
}
\value{
a tibble with three columns: "draw_index", "mean", "sd"
}
\description{
Sample reporting parameters using Metropolis MCMC
}
