% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{nb_log_likelihood_Rt_piece}
\alias{nb_log_likelihood_Rt_piece}
\title{Calculates negative-binomial log-likelhood within a single Rt piece}
\usage{
nb_log_likelihood_Rt_piece(Rt, kappa, w, onset_times, cases_df)
}
\arguments{
\item{Rt}{an Rt value for the piece}

\item{kappa}{overdispersion parameter}

\item{w}{weights corresponding to the generation times}

\item{onset_times}{the onset times corresponding to the piece}

\item{cases_df}{a tibble with 'cases_true' as a column which has been ordered
so that the latest onset times are at the bottom}
}
\value{
a log-likelihood of the piece
}
\description{
Calculates negative-binomial log-likelhood within a single Rt piece
}
