% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{observed_cases_single}
\alias{observed_cases_single}
\title{Cases arising on a given day which are reported between two subsequent days}
\usage{
observed_cases_single(
  cases_observed,
  cases_true,
  day_2,
  day_1,
  day_onset,
  reporting_parameters
)
}
\arguments{
\item{cases_observed}{observed case count for cases originating on day_onset on
day_1}

\item{cases_true}{true case count originating on day_onset}

\item{day_2}{day of last observation: day_2 > day_1 >= day_onset}

\item{day_1}{day of first observation: day_1 >= day_onset}

\item{day_onset}{day when case originates}

\item{reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}
}
\value{
a count representing number of cases reported between day_1 and day_2
}
\description{
The number of cases originating on day_onset which are subsequently reported
between day_1 and day_2 is modelled as a binomial distribution:
\deqn{cases_reported\sim binomial(cases_true-cases_observed, p_detect)}
where cases_observed is the number of cases originating on day_onset reported on
day_1; p_detect is the probability a case was undetected on day_1 but
is detected by day_2.
}
