% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{observed_cases_trajectory}
\alias{observed_cases_trajectory}
\title{Generate trajectory of reported cases for a given count originating on a
particular day}
\usage{
observed_cases_trajectory(
  cases_true,
  days_reporting,
  day_onset,
  reporting_parameters,
  is_gamma_delay
)
}
\arguments{
\item{cases_true}{true case count originating on day_onset}

\item{days_reporting}{a vector of days on which cases were counted}

\item{day_onset}{day when case originates}

\item{reporting_parameters}{named list of 'location' and 'scale' of distribution
characterising the reporting delay distribution}

\item{is_gamma_delay}{if TRUE, the reporting delay distribution is thought to
be a gamma; if FALSE, a lognormal distribution}
}
\value{
a vector of reported case counts of same length as days_reporting
}
\description{
Generate trajectory of reported cases for a given count originating on a
particular day
}
