% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pgamma_mean_sd}
\alias{pgamma_mean_sd}
\title{Reparameterisation of gamma cdf using mean and sd}
\usage{
pgamma_mean_sd(x, mu, sigma)
}
\arguments{
\item{x}{a value to be evaluated at}

\item{mu}{a mean}

\item{sigma}{a standard deviation}
}
\value{
a cumulative density
}
\description{
Reparameterisation of gamma cdf using mean and sd
}
