% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{prepare_stan_data_and_init}
\alias{prepare_stan_data_and_init}
\title{Title}
\usage{
prepare_stan_data_and_init(
  df,
  current_values,
  priors,
  is_negative_binomial,
  is_rw_prior,
  serial_parameters,
  serial_max,
  n_iterations,
  is_gamma_delay
)
}
\arguments{
\item{is_gamma_delay}{}
}
\value{

}
\description{
Title
}
