% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{prior_reporting_parameters}
\alias{prior_reporting_parameters}
\title{Gamma prior for reporting parameters}
\usage{
prior_reporting_parameters(current_reporting_parameters, prior_parameters)
}
\arguments{
\item{current_reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}

\item{prior_parameters}{named list with elements 'mean_mu', 'mean_sigma', 'sd_mu',
'sd_sigma' representing the gamma prior parameters for the mean and sd
parameters of the reporting parameters (itself described by a gamma
distribution)}
}
\value{
a log-probability density
}
\description{
Gamma prior for reporting parameters
}
