% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{propose_overdispersion_parameter}
\alias{propose_overdispersion_parameter}
\title{Propose new overdispersion parameter by sampling from a normal centered
on the current value}
\usage{
propose_overdispersion_parameter(
  overdispersion_current,
  metropolis_overdispersion_sd
)
}
\arguments{
\item{overdispersion_current}{an overdispersion parameter value (should exceed 0)}

\item{metropolis_overdispersion_sd}{the standard deviation of the proposal kernel}
}
\value{
a proposed overdispersion parameter value
}
\description{
Propose new overdispersion parameter by sampling from a normal centered
on the current value
}
