% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{propose_reporting_parameters}
\alias{propose_reporting_parameters}
\title{Propose new reporting parameters using normal kernel
centered at current values}
\usage{
propose_reporting_parameters(
  current_reporting_parameters,
  metropolis_parameters
)
}
\arguments{
\item{current_reporting_parameters}{a tibble with column names: "reporting_piece_index", "mean", "sd"}

\item{metropolis_parameters}{named list of 'mean_step', 'sd_step' containing
step sizes for Metropolis step}
}
\value{
a tibble with column names: "reporting_piece_index", "mean", "sd"
}
\description{
Propose new reporting parameters using normal kernel
centered at current values
}
