% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{qgamma_mean_sd}
\alias{qgamma_mean_sd}
\title{Reparameterisation of gamma inverse-cdf using mean and sd}
\usage{
qgamma_mean_sd(x, mu, sigma)
}
\arguments{
\item{x}{a value to be evaluated at}

\item{mu}{a mean}

\item{sigma}{a standard deviation}
}
\value{
a positive value
}
\description{
Reparameterisation of gamma inverse-cdf using mean and sd
}
