% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_or_maximise_gamma}
\alias{sample_or_maximise_gamma}
\title{Draws from the gamma distribution or returns the value which maximises
it}
\usage{
sample_or_maximise_gamma(shape, rate, ndraws, maximise = FALSE)
}
\arguments{
\item{shape}{the shape parameter of a gamma distribution}

\item{rate}{the rate parameter of a gamma distribution}

\item{ndraws}{number of draws if maximise=FALSE}

\item{maximise}{whether to return the mode of the gamma distribution}
}
\value{
a value or (if ndraws > 1) a vector of values
}
\description{
Draws from the gamma distribution or returns the value which maximises
it
}
