% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_reporting}
\alias{sample_reporting}
\title{Draw reporting parameter values either by sampling or by
maximising}
\usage{
sample_reporting(
  snapshot_with_true_cases_df,
  current_reporting_parameters,
  prior_parameters,
  metropolis_parameters,
  maximise = FALSE,
  ndraws = 1
)
}
\arguments{
\item{snapshot_with_true_cases_df}{a tibble with four columns:
time_onset, time_reported, cases_reported, cases_true}

\item{current_reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}

\item{prior_parameters}{named list with elements 'mean_mu', 'mean_sigma', 'sd_mu',
'sd_sigma' representing the gamma prior parameters for the mean and sd
parameters of the reporting parameters (itself described by a gamma
distribution)}

\item{metropolis_parameters}{named list of 'mean_step', 'sd_step' containing
step sizes for Metropolis step}

\item{maximise}{if true choose reporting parameters by maximising
log-probability; else (default) use Metropolis MCMC
to draw parameters}

\item{ndraws}{number of iterates of the Markov chain to simulate}
}
\value{
a tibble with three columns: "draw_index", "mean, "sd"
}
\description{
Draw reporting parameter values either by sampling or by
maximising
}
