% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_Rt}
\alias{sample_Rt}
\title{Sample piecewise-constant Rt values}
\usage{
sample_Rt(
  cases_history_df,
  Rt_prior_parameters,
  serial_parameters,
  kappa = NULL,
  serial_max = 40,
  ndraws = 1,
  maximise = FALSE
)
}
\arguments{
\item{cases_history_df}{a tibble with three columns: time_onset, cases_true
and Rt_index}

\item{Rt_prior_parameters}{a list with elements 'shape' and 'rate' describing
the gamma prior for Rt}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}

\item{kappa}{overdispersion parameter}

\item{serial_max}{maximum point at which to truncate sum in renewal process}

\item{ndraws}{number of draws of Rt}

\item{maximise}{rather than sample a case count give the case count with the
maximum probability (by default is FALSE)}
}
\value{
a tibble with three columns: "Rt_piece_index", "draw_index", "Rt"
}
\description{
If the renewal model is specified by a Poisson (the default):
\deqn{cases_true_t ~ Poisson(Rt * \sum_tau=1^t_max w_t cases_true_t-tau)}
If an Rt value is given a gamma prior, this results in a posterior
distribution:
\deqn{Rt ~ gamma(alpha + cases_true_t, beta + \sum_tau=1^t_max w_t cases_true_t-tau))}
where alpha and beta are the shape and rate parameters of the gamma
prior distribution. Here, we assume that Rt is constant over a set of onset
times 'onset_time_set'. This means that the posterior for a single Rt value is given
by:
\deqn{Rt ~ gamma(alpha + \sum_{t in onset_time_set} cases_true_t,
      beta + \sum_{t in onset_time_set}\sum_tau=1^t_max w_t cases_true_t-tau))}
This function either returns a draw (or draws if ndraws>1) from
this posterior, or it returns the Rt set that maximises it
(if maximise=TRUE).
Alternatively the renewal equation may be specified by a negative binomial distribution:
\deqn{cases_true_t ~ NB(Rt * \sum_tau=1^t_max w_t cases_true_t-tau, kappa)}
where kappa is the overdispersion parameter. In this case, importance sampling
using the Poisson posterior as the importance distribution is used to
estimate a negative binomial posterior.
}
