% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_Rt_single_piece}
\alias{sample_Rt_single_piece}
\title{Sample a single Rt value corresponding to a single piecewise-
constant element of an Rt vector}
\usage{
sample_Rt_single_piece(
  Rt_piece_index,
  cases_history_df,
  Rt_prior_parameters,
  serial_parameters,
  kappa = NULL,
  serial_max = 40,
  ndraws = 1,
  maximise = FALSE,
  is_negative_binomial = FALSE,
  nresamples = 100
)
}
\arguments{
\item{Rt_piece_index}{the index of the Rt piece being sampled}

\item{cases_history_df}{a tibble with three columns: time_onset, cases_true
and Rt_index}

\item{Rt_prior_parameters}{a list with elements 'shape' and 'rate' describing
the gamma prior for Rt}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}

\item{kappa}{overdispersion parameter}

\item{serial_max}{maximum point at which to truncate sum in renewal process}

\item{ndraws}{number of draws of Rt}

\item{maximise}{rather than sample a case count give the case count with the
maximum probability (by default is FALSE)}

\item{is_negative_binomial}{if negative-binomial renewal model specified (defaults to FALSE)}

\item{nresamples}{number of importance resamples of Rt to perform if assuming a
negative binomial model}
}
\value{
a draw (or draws) for Rt
}
\description{
Sample a single Rt value corresponding to a single piecewise-
constant element of an Rt vector
}
