% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_true_cases}
\alias{sample_true_cases}
\title{Samples a case count arising on a given onset day}
\usage{
sample_true_cases(
  cases_true,
  observation_df,
  cases_history,
  Rt,
  day_onset,
  serial_parameters,
  reporting_parameters,
  ndraws = 1
)
}
\arguments{
\item{cases_true}{a vector of possible case counts}

\item{observation_df}{a two-column data frame with columns 'day' and 'reported_cases'}

\item{cases_history}{a vector containing history of cases arranged from
recent to past}

\item{Rt}{effective reproduction number}

\item{day_onset}{day when case originates}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}

\item{reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}

\item{ndraws}{number of draws of cases}
}
\value{
a sampled case count arising on a given onset day
}
\description{
The distribution being drawn from is given by:
\deqn{p(cases_true_t|data, Rt, reporting_params, serial_params) \propto p(data|cases_true, reporting_params)
 p(cases_true_t|cases_true_t_1, cases_true_t_2, ..., Rt, serial_params)}
}
