% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{sample_true_cases_single_onset}
\alias{sample_true_cases_single_onset}
\title{Samples a case count arising on a given onset day}
\usage{
sample_true_cases_single_onset(
  observation_df,
  cases_history,
  max_cases,
  Rt,
  day_onset,
  serial_parameters,
  reporting_parameters,
  ndraws = 1,
  maximise = FALSE,
  kappa = NULL,
  is_negative_binomial = FALSE
)
}
\arguments{
\item{observation_df}{a two-column data frame with columns 'time_reported' and 'cases_reported'}

\item{cases_history}{a vector containing history of cases arranged from
recent to past}

\item{max_cases}{maximum possible cases thought to arise on a given day}

\item{Rt}{effective reproduction number}

\item{day_onset}{day when case originates}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}

\item{reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}

\item{ndraws}{number of draws of cases}

\item{maximise}{rather than sample a case count give the case count with the
maximum probability (by default is FALSE)}

\item{kappa}{overdispersion parameter}

\item{is_negative_binomial}{if negative-binomial renewal model specified (defaults to FALSE)}
}
\value{
a sampled case count arising on a given onset day
}
\description{
The distribution being drawn from is given by:
\deqn{p(cases_true_t|data, Rt, reporting_params, serial_params) \propto p(data|cases_true, reporting_params)
 p(cases_true_t|cases_true_t_1, cases_true_t_2, ..., Rt, serial_params)}
}
