% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{stan_initialisation}
\alias{stan_initialisation}
\title{Title}
\usage{
stan_initialisation(
  df,
  current_parameter_values,
  priors,
  is_negative_binomial,
  is_rw_prior,
  serial_parameters,
  serial_max,
  is_gamma_delay,
  stan_model
)
}
\arguments{
\item{stan_model}{}
}
\value{

}
\description{
Title
}
