% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{state_process_nb_logp_all_onsets}
\alias{state_process_nb_logp_all_onsets}
\title{Calculates the overall log-probability of cases conditional on their history
and Rt values, assuming a negative-binomial renewal model}
\usage{
state_process_nb_logp_all_onsets(kappa, cases_history_rt_df, serial_parameters)
}
\arguments{
\item{kappa}{an overdispersion parameter value (if <= 0, returns -Inf)}

\item{cases_history_rt_df}{a tibble with three columns: 'time_onset',
'cases_true' and 'Rt'}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}
}
\value{
a log-probability value
}
\description{
Specifically, this assumes the probability is of the form:
\deqn{\prod_tau=1^T NB(cases_true_tau|Rt * \sum_s=1^tau-1 w_s cases_true_tau-s, kappa)}
}
