% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{thin_series}
\alias{thin_series}
\title{Thins cases}
\usage{
thin_series(case_obs)
}
\arguments{
\item{case_obs}{a tibble produced by observed_cases}
}
\value{
a thinned tibble
}
\description{
Only those data points around where the case count changes contain
information. This method searches and keeps only those data points.
}
