% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{thin_single_series}
\alias{thin_single_series}
\title{Thins cases for a single onset time}
\usage{
thin_single_series(case_obs_single)
}
\arguments{
\item{case_obs_single}{a tibble produced by observed_cases but subsetted to
a single onset time}
}
\value{
a thinned tibble
}
\description{
Only those data points around where the case count changes contain
information. This method searches and keeps only those data points.
}
