% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{true_cases}
\alias{true_cases}
\title{Generate true case series}
\usage{
true_cases(
  days_total,
  Rt_function,
  kappa,
  serial_parameters,
  initial_parameters = list(mean = 3, length = 5),
  serial_max = 40
)
}
\arguments{
\item{days_total}{number of days to simulate process for (post-initial
period)}

\item{Rt_function}{takes day as an input and outputs an Rt value}

\item{kappa}{over-dispersion parameter of negative binomial}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}

\item{initial_parameters}{a named list of 'mean' and 'length'
which is used to generate seed cases by sampling from a negative binomial
distribution}

\item{serial_max}{maximum point at which to truncate sum in renewal process}
}
\value{
vector of true cases of length days_total
}
\description{
Cases are assumed to be generated by a negative-binomial renewal
process:
\deqn{cases_t \sim neg-binomial(Rt \sum_s=1^\infty w(s) cases_t-s, kappa)}
where kappa is the over-dispersion parameter. Here, we truncate this
sum at a user specified lag (determined by the length of the weight vector).
}
\examples{
library(incidenceinflation)
# generate case series for 100 days with time-varying Rt
days_total <- 100

# make Rt interpolation function
v_Rt <- c(rep(1.3, 25), rep(1, 25), rep(2, 50))
Rt_function <- stats::approxfun(1:days_total, v_Rt)

# serial interval parameters
s_params <- list(mean=5, sd=1)

# renewal parameter
kappa <- 2
cases <- true_cases(days_total, Rt_function, kappa, s_params)
}
