% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{true_cases_single}
\alias{true_cases_single}
\title{Generates true cases for a single day}
\usage{
true_cases_single(R0, kappa, cases_history, weights)
}
\arguments{
\item{R0}{reproduction number}

\item{kappa}{over-dispersion parameter of negative binomial}

\item{cases_history}{a vector containing history of cases arranged from
recent to past}

\item{weights}{a vector containing weights representing the latency
distribution}
}
\value{
a count of cases
}
\description{
Cases are assumed to be generated by a negative-binomial renewal
process:
\deqn{cases_t \sim neg-binomial(R0 \sum_s=1^\infty w(s) cases_t-s, kappa)}
where kappa is the over-dispersion parameter. Here, we truncate this
sum at a user specified lag (determined by the length of the weight vector).
}
