% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{undetected_prob}
\alias{undetected_prob}
\title{Probability case remains undetected over time}
\usage{
undetected_prob(t_1, t_onset, reporting_parameters, is_gamma_delay)
}
\arguments{
\item{t_1}{observation time subsequent to time when case originates: t1>=t0}

\item{t_onset}{time when case originates}

\item{reporting_parameters}{named list of 'location' and 'scale' of distribution
characterising the reporting delay distribution}

\item{is_gamma_delay}{if TRUE, the reporting delay distribution is thought to
be a gamma; if FALSE, a lognormal distribution}
}
\value{
a probability
}
\description{
This is effectively a survival function, which shows the
probability that an actual case originating on day t_onset
remains unreported at t_1.
}
\examples{
incidenceinflation:::undetected_prob(2, 1, list(location=10, scale=5))
}
