% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{combine_chains}
\alias{combine_chains}
\title{Combines Markov chains across multiple runs of mcmc_single}
\usage{
combine_chains(list_of_results, is_negative_binomial = FALSE)
}
\arguments{
\item{list_of_results}{a list of results, where each element is a result of running mcmc_single}

\item{is_negative_binomial}{if negative-binomial renewal model specified (defaults to FALSE)}
}
\value{
a named list of three tibbles: "cases", "Rt" and "reporting" (and if running
a negative binomial model an additional 'overdispersion' element) which
contain estimates of the model parameters with chain index  included
}
\description{
Combines Markov chains across multiple runs of mcmc_single
}
