% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{conditional_cases_logp}
\alias{conditional_cases_logp}
\title{Calculates (unnormalised) log-probability of a true case count pertaining to cases arising
on a particular onset day}
\usage{
conditional_cases_logp(
  cases_true,
  observation_df,
  cases_history,
  Rt,
  day_onset,
  serial_parameters,
  reporting_parameters
)
}
\arguments{
\item{cases_true}{true case count originating on day_onset}

\item{observation_df}{a two-column data frame with columns 'day' and 'reported_cases'}

\item{cases_history}{a vector containing history of cases arranged from
recent to past}

\item{Rt}{effective reproduction number}

\item{day_onset}{day when case originates}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}

\item{reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}
}
\value{
an (unnormalised) log-probability or vector of such log-probabilities
}
\description{
There are two components to this:
\deqn{p(cases_true_t|data, Rt, reporting_params, serial_params) \propto p(data|cases_true, reporting_params)
 p(cases_true_t|cases_true_t_1, cases_true_t_2, ..., Rt, serial_params)}
}
