% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{maximise_reporting_logp}
\alias{maximise_reporting_logp}
\title{Select reporting parameters by maximising log-probability}
\usage{
maximise_reporting_logp(
  snapshot_with_true_cases_df,
  current_reporting_parameters
)
}
\arguments{
\item{snapshot_with_true_cases_df}{a tibble with four columns:
time_onset, time_reported, cases_reported, cases_true}

\item{current_reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}
}
\value{
a tibble with three columns: "draw_index", "mean, "sd"
}
\description{
Select reporting parameters by maximising log-probability
}
