% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{mcmc}
\alias{mcmc}
\title{Runs MCMC or optimisation to estimate Rt, cases and reporting parameters}
\usage{
mcmc(
  niterations,
  data,
  priors,
  serial_parameters,
  initial_cases_true,
  initial_reporting_parameters,
  initial_Rt,
  stan_model,
  nchains = 1,
  initial_overdispersion = NULL,
  initial_sigma = NULL,
  is_negative_binomial = FALSE,
  is_gamma_delay = TRUE,
  serial_max = 40,
  p_gamma_cutoff = 0.99,
  maximise = FALSE,
  print_to_screen = TRUE,
  is_parallel = FALSE
)
}
\arguments{
\item{niterations}{number of MCMC iterations to run or number of iterative maximisations to run}

\item{data}{a tibble with five columns: time_onset, time_reported, cases_reported, Rt_index, reporting_index}

\item{priors}{a named list with: 'Rt', 'reporting', 'max_cases'. These take
the form: 'Rt' is a named list with elements 'shape' and 'rate' describing
the gamma prior for each Rt; 'reporting' is a named list with elements
'mean_mu', 'mean_sigma', 'sd_mu', 'sd_sigma' representing the gamma
prior parameters for the mean and sd parameters of the reporting parameters
(itself described by a gamma distribution); max cases controls the upper
limit of the discrete uniform distribution representing the prior on true
cases}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}

\item{initial_cases_true}{a tibble with two columns: "time_onset" and "cases_true", which represents initial
estimates of the true number of cases with each onset time.}

\item{initial_reporting_parameters}{provides initial guesses of the mean and
sd of the reporting delay distribution(s). These can be either a named with two named elements
('mean', 'sd') for a time-invariant reporting delay or a tibble with three columns:
'reporting_piece_index', 'mean', 'sd' (where the number of indices corresponds to the number
provided in the data frame).}

\item{initial_Rt}{initial guess of the Rt values in each of the piecewise segments.
Provided in the form of a tibble with columns: 'Rt_index' and 'Rt'}

\item{nchains}{number of Markov chains to run. Defaults to 1}

\item{initial_overdispersion}{the initial value of the overdispersion parameter if
assuming a negative binomial sampling model (default to 5)}

\item{initial_sigma}{the initial value of the sigma parameter that characterises
the random walk prior if selected for Rt}

\item{is_negative_binomial}{Boolean indicating whether to use a negative binomial renewal model
or a Poisson model (if FALSE). By default a Poisson model is assumed}

\item{is_gamma_delay}{if TRUE, the reporting delay distribution is thought to
be a gamma; if FALSE, a lognormal distribution}

\item{serial_max}{maximum point at which to truncate sum in renewal process}

\item{p_gamma_cutoff}{a p value (0 <= p <= 1) indicating the threshold above which
we deem certainty}

\item{maximise}{rather than sample a case count give the case count with the
maximum probability (by default is FALSE)}

\item{print_to_screen}{prints progress of MCMC sampling to screen. Defaults to true. Disabled when is_parallel is TRUE.}

\item{is_parallel}{Boolean to indicate whether or not to run chains in parallel. Defaults to FALSE.}
}
\value{
a named list of three tibbles: "cases", "Rt" and "reporting" which contain estimates of the model parameters
}
\description{
Runs MCMC or optimisation to estimate Rt, cases and reporting parameters
}
