% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{mcmc_single}
\alias{mcmc_single}
\title{Runs MCMC or optimisation to estimate Rt, cases and reporting parameters}
\usage{
mcmc_single(
  niterations,
  snapshot_with_Rt_index_df,
  priors,
  serial_parameters,
  initial_cases_true,
  initial_reporting_parameters,
  initial_Rt,
  reporting_metropolis_parameters = list(mean_step = 0.25, sd_step = 0.1),
  serial_max = 40,
  p_gamma_cutoff = 0.99,
  maximise = FALSE,
  print_to_screen = TRUE
)
}
\arguments{
\item{niterations}{number of MCMC iterations to run or number of iterative maximisations to run}

\item{snapshot_with_Rt_index_df}{a tibble with four columns: time_onset, time_reported,
cases_reported, Rt_index; an optional fifth column can be provided named reporting_index,
which specifies which reporting distribubtion each onset time corresponds to}

\item{priors}{a named list with: 'Rt', 'reporting', 'max_cases'. These take
the form: 'Rt' is a named list with elements 'shape' and 'rate' describing
the gamma prior for each Rt; 'reporting' is a named list with elements
'mean_mu', 'mean_sigma', 'sd_mu', 'sd_sigma' representing the gamma
prior parameters for the mean and sd parameters of the reporting parameters
(itself described by a gamma distribution); max cases controls the upper
limit of the discrete uniform distribution representing the prior on true
cases}

\item{serial_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the serial interval distribution}

\item{initial_cases_true}{a tibble with two columns: "time_onset" and "cases_true", which represents initial
estimates of the true number of cases with each onset time.}

\item{initial_reporting_parameters}{provides initial guesses of the mean and
sd of the reporting delay distribution(s). These can be either a named with two named elements
('mean', 'sd') for a time-invariant reporting delay or a tibble with three columns:
'reporting_index', 'mean', 'sd' (where the number of indices corresponds to the number
provided in the data frame).}

\item{initial_Rt}{initial guess of the Rt values in each of the piecewise segments.
Provided in the form of a tibble with columns: 'Rt_index' and 'Rt'}

\item{reporting_metropolis_parameters}{named list of 'mean_step', 'sd_step' containing
step sizes for Metropolis step}

\item{serial_max}{maximum point at which to truncate sum in renewal process}

\item{p_gamma_cutoff}{a p value (0 <= p <= 1) indicating the threshold above which
we deem certainty}

\item{maximise}{whether to estimate MAP values of parameters (if true) or
sample parameter values using MCMC (if false). By default this is false.}

\item{print_to_screen}{prints progress of MCMC sampling to screen. Defaults to true.}
}
\value{
a named list of three tibbles: "cases", "Rt" and "reporting" which contain estimates of the model parameters
}
\description{
Runs MCMC or optimisation to estimate Rt, cases and reporting parameters
}
