% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{metropolis_step}
\alias{metropolis_step}
\title{Sample reporting parameters using a single Metropolis step}
\usage{
metropolis_step(
  snapshot_with_true_cases_df,
  current_reporting_parameters,
  logp_current,
  prior_parameters,
  metropolis_parameters
)
}
\arguments{
\item{snapshot_with_true_cases_df}{a tibble with four columns:
time_onset, time_reported, cases_reported, cases_true}

\item{current_reporting_parameters}{a tibble with column names: "reporting_piece_index", "mean", "sd"}

\item{prior_parameters}{named list with elements 'mean_mu', 'mean_sigma', 'sd_mu',
'sd_sigma' representing the gamma prior parameters for the mean and sd
parameters of the reporting parameters (itself described by a gamma
distribution)}

\item{metropolis_parameters}{named list of 'mean_step', 'sd_step' containing
step sizes for Metropolis step}
}
\value{
a named list with two elements: 'parameter', a parameter value (which
may be a non-scalar); and 'logp', the log-probability corresponding to the
parameter value returned
}
\description{
Sample reporting parameters using a single Metropolis step
}
