% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{metropolis_steps}
\alias{metropolis_steps}
\title{Sample reporting parameters using Metropolis MCMC}
\usage{
metropolis_steps(
  snapshot_with_true_cases_df,
  current_reporting_parameters,
  prior_parameters,
  metropolis_parameters,
  ndraws
)
}
\arguments{
\item{snapshot_with_true_cases_df}{a tibble with four columns:
time_onset, time_reported, cases_reported, cases_true}

\item{current_reporting_parameters}{a tibble with column names: "reporting_piece_index", "mean", "sd"}

\item{prior_parameters}{named list with elements 'mean_mu', 'mean_sigma', 'sd_mu',
'sd_sigma' representing the gamma prior parameters for the mean and sd
parameters of the reporting parameters (itself described by a gamma
distribution)}

\item{metropolis_parameters}{named list of 'mean_step', 'sd_step' containing
step sizes for Metropolis step}

\item{ndraws}{number of iterates of the Markov chain to simulate}
}
\value{
a tibble with four columns: "reporting_piece_index", "draw_index", "mean", "sd"
}
\description{
Sample reporting parameters using Metropolis MCMC
}
