% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{observation_process_all_times_logp}
\alias{observation_process_all_times_logp}
\title{Observation probability across all onset times}
\usage{
observation_process_all_times_logp(
  snapshot_with_true_cases_df,
  reporting_parameters,
  is_gamma_delay = TRUE
)
}
\arguments{
\item{snapshot_with_true_cases_df}{a tibble with four columns:
time_onset, time_reported, cases_reported, cases_true}

\item{reporting_parameters}{a tibble with columns: 'reporting_piece_index', 'mean', 'sd'}
}
\value{
a log probability
}
\description{
Observation probability across all onset times
}
