% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{observation_process_all_times_logp}
\alias{observation_process_all_times_logp}
\title{Observation probability across all onset times}
\usage{
observation_process_all_times_logp(
  snapshot_with_true_cases_df,
  reporting_parameters
)
}
\arguments{
\item{snapshot_with_true_cases_df}{a tibble with four columns:
time_onset, time_reported, cases_reported, cases_true}

\item{reporting_parameters}{either a named list of 'mean' and 'sd' of gamma
distribution indicating a single reporting delay distribution for the whole period
or a tibble with columns: 'time_onset', 'mean', 'sd' indicating the reporting
delay distribution for each time period}
}
\value{
a log probability
}
\description{
Observation probability across all onset times
}
