% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probabilities.R
\name{observation_process_logp}
\alias{observation_process_logp}
\title{Calculates observation process log probability density}
\usage{
observation_process_logp(
  observation_df,
  cases_true,
  day_onset,
  reporting_parameters,
  is_gamma_delay
)
}
\arguments{
\item{observation_df}{a two-column data frame with columns 'time_reported' and 'cases_reported'}

\item{cases_true}{true case count(s) originating on day_onset: can be a single value or vector}

\item{day_onset}{day when case originates}

\item{reporting_parameters}{named list of 'location' and 'scale' of distribution
characterising the reporting delay distribution}

\item{is_gamma_delay}{if TRUE, the reporting delay distribution is thought to
be a gamma; if FALSE, a lognormal distribution}
}
\value{
a log-probability or vector of log-probabilities
}
\description{
The probability of observations is a function of the number of cases remaining
to be reported:
\deqn{I_remaining = I_true - I_day_1}
and the number of cases observed between day_1 and day_2:
\deqn{I_obs = I_day_2 - I_day_1}
It is given by:
\deqn{binomial(I_obs|I_remaining, p_detect)}
where p_detect is the probability a thus undetected case is detected between
day_1 and day_2.
This function calculates the combined probability of all observed cases pertaining
to a particular onset date.
}
