% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{observed_cases}
\alias{observed_cases}
\title{Generate reported case trajectories for each day when cases appear}
\usage{
observed_cases(I_true, reporting_parameters, days_max_follow_up = 30)
}
\arguments{
\item{I_true}{a vector of true cases originating each day}

\item{reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}

\item{days_max_follow_up}{max days at which to simulate reporting case
trajectory}
}
\value{
a tibble with observed case trajectories for each time of onset
}
\description{
Generate reported case trajectories for each day when cases appear
}
\examples{
library(incidenceinflation)
observed_cases(stats::rpois(5, 5), list(mean=5, sd=1))
}
