% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{observed_cases}
\alias{observed_cases}
\title{Generate reported case trajectories for each day when cases appear}
\usage{
observed_cases(cases_true, reporting_parameters, days_max_follow_up = 30)
}
\arguments{
\item{cases_true}{a vector of true cases originating each day}

\item{reporting_parameters}{either a named list of 'mean' and 'sd' of gamma
distribution indicating a single reporting delay distribution for the whole period
or a tibble with columns: 'time_onset', 'mean', 'sd' indicating the reporting
delay distribution for each time period}

\item{days_max_follow_up}{max days at which to simulate reporting case
trajectory}
}
\value{
a tibble with observed case trajectories for each time of onset
}
\description{
Generate reported case trajectories for each day when cases appear
}
\examples{
library(incidenceinflation)
observed_cases(stats::rpois(5, 5), list(mean=5, sd=1))
}
