% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{observed_cases_trajectory}
\alias{observed_cases_trajectory}
\title{Generate trajectory of reported cases for a given count originating on a
particular day}
\usage{
observed_cases_trajectory(
  cases_true,
  days_reporting,
  day_onset,
  reporting_parameters
)
}
\arguments{
\item{cases_true}{true case count originating on day_onset}

\item{days_reporting}{a vector of days on which cases were counted}

\item{day_onset}{day when case originates}

\item{reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}
}
\value{
a vector of reported case counts of same length as days_reporting
}
\description{
Generate trajectory of reported cases for a given count originating on a
particular day
}
