% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{propose_reporting_parameters}
\alias{propose_reporting_parameters}
\title{Propose new reporting parameters using normal kernel
centered at current values}
\usage{
propose_reporting_parameters(
  current_reporting_parameters,
  metropolis_parameters
)
}
\arguments{
\item{current_reporting_parameters}{named list of 'mean' and 'sd' of gamma distribution
characterising the reporting delay distribution}

\item{metropolis_parameters}{named list of 'mean_step', 'sd_step' containing
step sizes for Metropolis step}
}
\value{
list of reporting parameters
}
\description{
Propose new reporting parameters using normal kernel
centered at current values
}
