% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{stan_initialisation}
\alias{stan_initialisation}
\title{Initialises Stan model to allow access to log_prob of model}
\usage{
stan_initialisation(
  df,
  current_values,
  priors,
  is_negative_binomial,
  is_rw_prior,
  serial_parameters,
  serial_max,
  is_gamma_delay,
  stan_model
)
}
\arguments{
\item{is_gamma_delay}{}

\item{stan_model}{an object generated by running cmdstanr::cmdstan_model}
}
\value{
an initialised model
}
\description{
Needs to be rerun each time any data being passed to the Stan file changes
}
