% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated-data.R
\name{undetected_prob}
\alias{undetected_prob}
\title{Probability case remains undetected over time}
\usage{
undetected_prob(t1, t0, delay_mean, delay_sd)
}
\arguments{
\item{t1}{time subsequent to time when case originates, t1>=t0}

\item{t0}{time when case originates}

\item{delay_mean}{mean of gamma distribution governing reporting delay distribution}

\item{delay_sd}{std. dev. of gamma distribution governing reporting delay distribution}
}
\value{
a probability
}
\description{
This is effectively a survival function, which shows the
probability that an actual case originating on day t0
remains unreported at t1.
}
